ST-917 Event Names moved to a constant class
