+++ src/integrationTest/java/uk/gov/hmcts/sptribs/caseworker/CaseworkerIssueCaseIT.java
@@ -23,7 +23,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import static uk.gov.hmcts.sptribs.caseworker.event.CaseworkerIssueCase.CASEWORKER_ISSUE_CASE;
+import static uk.gov.hmcts.sptribs.testutil.ITEventConstants.CASEWORKER_ISSUE_CASE;
 import static uk.gov.hmcts.sptribs.testutil.IdamWireMock.COURT_ADMIN_CIC;
 import static uk.gov.hmcts.sptribs.testutil.IdamWireMock.stubForIdamDetails;
 import static uk.gov.hmcts.sptribs.testutil.TestConstants.ABOUT_TO_SUBMIT_URL;
