+++ src/integrationTest/java/uk/gov/hmcts/sptribs/testutil/ITEventConstants.java
@@ -1,0 +1,11 @@
+package uk.gov.hmcts.sptribs.testutil;
+
+public final class ITEventConstants {
+
+    public static final String CASEWORKER_ADD_NOTE = "caseworker-add-note";
+    public static final String CASEWORKER_ISSUE_CASE = "caseworker-issue-case";
+
+    private ITEventConstants() {
+    }
+
+}
