+++ src/main/java/uk/gov/hmcts/sptribs/caseworker/event/CaseworkerAmendCase.java
@@ -10,6 +10,7 @@ import uk.gov.hmcts.sptribs.ciccase.model.UserRole;
 import uk.gov.hmcts.sptribs.common.ccd.CcdPageConfiguration;
 import uk.gov.hmcts.sptribs.common.ccd.PageBuilder;
 
+import static uk.gov.hmcts.sptribs.caseworker.util.EventConstants.CASEWORKER_AMEND_CASE;
 import static uk.gov.hmcts.sptribs.ciccase.model.State.POST_SUBMISSION_STATES;
 import static uk.gov.hmcts.sptribs.ciccase.model.UserRole.COURT_ADMIN_CIC;
 import static uk.gov.hmcts.sptribs.ciccase.model.UserRole.SUPER_USER;
@@ -17,7 +18,6 @@ import static uk.gov.hmcts.sptribs.ciccase.model.access.Permissions.CREATE_READ_
 
 @Component
 public class CaseworkerAmendCase implements CCDConfig<CaseData, State, UserRole> {
-    public static final String CASEWORKER_AMEND_CASE = "caseworker-amend-case";
     private final CcdPageConfiguration amendCase = new AmendCase();
 
     @Override
