+++ src/main/java/uk/gov/hmcts/sptribs/caseworker/event/CaseworkerCloseTheCase.java
@@ -14,6 +14,7 @@ import uk.gov.hmcts.sptribs.ciccase.model.UserRole;
 import uk.gov.hmcts.sptribs.common.ccd.CcdPageConfiguration;
 import uk.gov.hmcts.sptribs.common.ccd.PageBuilder;
 
+import static uk.gov.hmcts.sptribs.caseworker.util.EventConstants.CASEWORKER_CLOSE_THE_CASE;
 import static uk.gov.hmcts.sptribs.ciccase.model.State.CaseClosed;
 import static uk.gov.hmcts.sptribs.ciccase.model.State.CaseManagement;
 import static uk.gov.hmcts.sptribs.ciccase.model.UserRole.COURT_ADMIN_CIC;
@@ -24,7 +25,6 @@ import static uk.gov.hmcts.sptribs.ciccase.model.access.Permissions.CREATE_READ_
 @Component
 @Slf4j
 public class CaseworkerCloseTheCase implements CCDConfig<CaseData, State, UserRole> {
-    public static final String CASEWORKER_CLOSE_THE_CASE = "caseworker-close-the-case";
 
     private static final CcdPageConfiguration closeCaseWarning = new CloseCaseWarning();
