+++ src/main/java/uk/gov/hmcts/sptribs/caseworker/event/CaseworkerRemoveStay.java
@@ -17,6 +17,7 @@ import uk.gov.hmcts.sptribs.common.ccd.CcdPageConfiguration;
 import uk.gov.hmcts.sptribs.common.ccd.PageBuilder;
 import uk.gov.hmcts.sptribs.common.notification.CaseUnstayedNotification;
 
+import static uk.gov.hmcts.sptribs.caseworker.util.EventConstants.CASEWORKER_REMOVE_STAY;
 import static uk.gov.hmcts.sptribs.ciccase.model.State.CaseStayed;
 import static uk.gov.hmcts.sptribs.ciccase.model.UserRole.COURT_ADMIN_CIC;
 import static uk.gov.hmcts.sptribs.ciccase.model.UserRole.SOLICITOR;
@@ -26,7 +27,6 @@ import static uk.gov.hmcts.sptribs.ciccase.model.access.Permissions.CREATE_READ_
 @Component
 @Slf4j
 public class CaseworkerRemoveStay implements CCDConfig<CaseData, State, UserRole> {
-    public static final String CASEWORKER_REMOVE_STAY = "caseworker-remove-stay";
 
     private static final CcdPageConfiguration removeStay = new RemoveStay();
