+++ src/main/java/uk/gov/hmcts/sptribs/caseworker/event/CaseworkerStayTheCase.java
@@ -16,6 +16,7 @@ import uk.gov.hmcts.sptribs.ciccase.model.UserRole;
 import uk.gov.hmcts.sptribs.common.ccd.PageBuilder;
 import uk.gov.hmcts.sptribs.common.notification.CaseStayedNotification;
 
+import static uk.gov.hmcts.sptribs.caseworker.util.EventConstants.CASEWORKER_STAY_THE_CASE;
 import static uk.gov.hmcts.sptribs.ciccase.model.State.CaseManagement;
 import static uk.gov.hmcts.sptribs.ciccase.model.State.CaseStayed;
 import static uk.gov.hmcts.sptribs.ciccase.model.UserRole.COURT_ADMIN_CIC;
@@ -26,7 +27,6 @@ import static uk.gov.hmcts.sptribs.ciccase.model.access.Permissions.CREATE_READ_
 @Component
 @Slf4j
 public class CaseworkerStayTheCase implements CCDConfig<CaseData, State, UserRole> {
-    public static final String CASEWORKER_STAY_THE_CASE = "caseworker-stay-the-case";
 
     @Autowired
     private CaseStayedNotification caseStayedNotification;
