+++ src/test/java/uk/gov/hmcts/sptribs/caseworker/event/CaseworkerManageCaseFlagTest.java
@@ -11,9 +11,9 @@ import uk.gov.hmcts.sptribs.ciccase.model.State;
 import uk.gov.hmcts.sptribs.ciccase.model.UserRole;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static uk.gov.hmcts.sptribs.caseworker.event.CaseworkerManageCaseFlag.CASEWORKER_MANAGE_CASE_FLAG;
 import static uk.gov.hmcts.sptribs.testutil.ConfigTestUtil.createCaseDataConfigBuilder;
 import static uk.gov.hmcts.sptribs.testutil.ConfigTestUtil.getEventsFrom;
+import static uk.gov.hmcts.sptribs.testutil.TestEventConstants.CASEWORKER_MANAGE_CASE_FLAG;
 
 @ExtendWith(MockitoExtension.class)
 class CaseworkerManageCaseFlagTest {
